requirements.txt
# 基础科学计算和数据处理库
numpy==2.0.0
pandas==2.2.2
scipy==1.15.1

# 数据获取和分析库
akshare==1.15.87
tqdm==4.67.1

# GUI库
PyQt6==6.8.1
markdown2==2.5.3

# 网络和API请求
requests==2.32.3
python-dotenv==1.0.1

# Web API服务器
fastapi==0.112.0
uvicorn==0.30.1
pydantic==2.7.1

# 跨域资源共享
starlette>=0.37.2,<0.38.0

# 日志和系统工具
loguru==0.7.2

# 可选：数据可视化（未来扩展）
matplotlib==3.9.2
seaborn==0.13.2

# 开发和调试工具
ipython>=7.34.0

# 其他依赖
beautifulsoup4==4.12.3
html5lib==1.1
lxml==4.9.4
jsonpath==0.82.2
openpyxl==3.1.5
